PHEE-345: fix NPE when converting Long object to int (#77)

+++ src/main/java/org/mifos/processor/bulk/zeebe/worker/BatchStatusWorker.java
@@ -34,7 +34,7 @@ public class BatchStatusWorker extends BaseWorker {
 
             Boolean batchStatusFailed = exchange.getProperty(BATCH_STATUS_FAILED, Boolean.class);
             if (batchStatusFailed == null || !batchStatusFailed) {
-                successRate = exchange.getProperty(COMPLETION_RATE, Long.class).intValue();
+                successRate = exchange.getProperty(COMPLETION_RATE, Integer.class);
             } else {
                 variables.put(ERROR_CODE, exchange.getProperty(ERROR_CODE));
                 variables.put(ERROR_DESCRIPTION, exchange.getProperty(ERROR_DESCRIPTION));
